Time: 0 ms (100.00%) | Memory: 103.6 MB (0.00%) - LeetSync

+++ 4005-maximum-total-subarray-value-i/maximum-total-subarray-value-i.cpp
@@ -4,7 +4,7 @@ public:
         int mx=*max_element(nums.begin(), nums.end());
         int mn= *min_element(nums.begin(), nums.end());
         long long diff= mx-mn;
-        long long ans= static_cast<long long>(k)* diff;
+        long long ans= (k)* diff;
         return ans;
     }
 };
